feat: Begin work on (pre|post)fix comments
We use an underscore before the key names to remain forwards-compatible
with the specification (it's still in draft, and it would be horrible [but
unlikely] if they added a 'Preifx-Comment' directive! It doesn't quite
seem to work -- there appear to be double newlines. Maybe it's the fault
of a previous commit.