style: Disable security test temporarily
Normally, testing for security issues is very important. However, in this
particular instance, on this particular line, we have determined that
there is no threat to security. Hence, we disable the check.

We have come to this conclusion by understanding that the relevant
security issue occurs when we allow unfiltered user input as the key to an
object. As JavaScript traverses the prototype chain, certain unexpected
values could be accessed.

This was a non-issue because we do a .hasOwnProperty check on the key we
intend to use. This means that the prototype chain will not be traversed
because a match will be found instantly (we use 'continue' to escape).

Additionally, the keys we attempt to use are already safe: we pick them
from a whitelist mapped with a pure function. When we apply this pure
function to each element in our whitelist, we don't (to the best of our
knowledge) get anything which may be harmful, such as 'constructor'. This
may become an issue in future if a 'Constructor' directive is added, but
given the previous defence this is still a non-issue.

When disabling lint, we do not perform a blanket disable. Instead, we
target a specific line for this specific security issue. If a different
security issue is introduced in future on that line, or this security
issue is introduced on a different line, it will be detected, hence this
change does not compromise security.