feat(#validatePolicyFields): Relax validation to allow objects; etc
We re-write the validation phase (still using Joi) such that objects can now be included. They are
not implemented yet.  In addition, we add (and implement!) _prefixComment and _postfixComment. We
use underscores to ensure this API is forwards-compatible with the specification (if, in the future,
a Postfix-Comment: directive were added).