test: Add test for comments (prefix, postfix, objects)
Some of these tests are intended to fail for the moment, this is part of TDD. We test that comment
objects (i.e. objects containing a 'value' key and optionally a 'comment' key), prefix comments
(comments at the head of a security.txt file), and postfix comments (comments at the tail thereof)
all work.