feat: Allow comments to be arrays
Whereas #formatSecurityPolicy could handle an array of comments (which would then be glued together
over multiple lines -- prefixed with the hash symbol to be compliant with the specification), the
validation schema was not aware of this: and required comments be strings. We also add tests to
ensure comments can always be arrays. Note: it might be good to factor out the schema for what a
'comment' is, so we only need to make a modification in one place in future.