test(formatPolicy.test.js): Test camelCase utility method to increase coverage
When converting a string to camelCase, there's a special case: the first word must be in lowercase,
but subsequent words must be capitalised. The specification currently only includes single words,
but there are proposals for multiple words seperated by dashes (and we support this if these are
accepted). The other tests therefore were only hitting the special case, and so part of the ternary
condition is untested. Coverage doesn't like this, so we add three tests for non-existant
directives: Abc, Abc-Def, Abc-Def-Ghi; and these restore coverage to 100%