fix(#validatePolicyFields): Disallow nesting of arrays through [{ value: []}]
#formatPolicyFields can't handle nested arrays. It's debatable if we should support nested arrays,
but the workaround (shoving the additional values outside the object but inside the array) doesn't
change the character count and seems reasonable to me. If we do want to support it, do we want to
support infinite nesting via this technique?

BREAKING CHANGE: Nested arrays using [{value: []}] will throw errors now