feat(#validatePolicyFields): Relax validation to prepare for deprecation
We add an object to store the deprecated fields, and what they should be. We then programatically
add the contents of this object to the schema. It could be prettier if we hard-code the deprecations
in the schema, but I figure we should keep them at the top of the file for maintainability