test(formatPolicy.test.js): Mock console.warn in tests
The test for deprecated keys would trigger a console.warning, per design, so that if a user was
using the deprecated spelling -- they would be told about it so that they could correct the issue.
However, the warning would be confusing during testing -- because the test was running correctly.
Hence, we mock console.warn so that the warning does not show up in the tests, and also ensures that
the warning is not accidentally removed.

#46