feat: Add the 'Preferred-Languages' directive
The Preferred-Languages directive allows security analysts to specify which languages they are most
comfortable conversing in when talking to security researchers about security reports. The directive
only allows the field to appear once in the document, and so if an array is passed it is
concatenated by commas into a string. Tests have been added to cover these cases. The validator does
not check whether the languages are compliant language codes. The tests do not verify that an object
for comments will be accepted, but it should because the validation schema uses the helper function
fieldValue().