fix(#validatePolicyFields): Disallow Array<object> in preferredLanguages
For other fields, if an array is allowed as a value, it means that the entire field will occur
multiple times within the security policy. Where that is the case, it makes sense for comments to be
prefixed before each occurence, and hence each array element should be allowed to be an object
(since those represent comment + value pairs). However, the preferredLanguages field uses arrays for
a different purpose: the array will eventually be concatenated into a comma-delimited string. As
such, it does not make sense for each array element to have a comment attached to it, because the
field will only ever occur once. Hence, disallow comment objects within an array of values, and test
for this also.