feat: Add and remove certain directives (#56)
Remove the Signature: and Permission: directives as the -05 draft removes these. Also, implement the Canonical: directive as the -05 draft adds it.

BREAKING CHANGE: Signature: and Permission: directives are both removed per the -05 version of the Internet draft. Using the `signature` or `permission` keys in an options object will lead to a validation error.

#52