feat: Redirect from /security.txt to /.well-known/security.txt
The language in the specification recently changed. Now, unless there are technical reasons for
which the /.well-known/ path cannot be used, sites should redirect from /security.txt to
/.well-known/security.txt. Further, this commit puts "/security.txt" into an array (of one item for
now), in case the specification adds more alternative URLs. The /.well-known path is moved into a
constant because it is now referenced twice (once for the redirection, and once for detecting if the
URL is one which demands a security policy). Tests have been modified to take full advantage of
Jest's mock functions (i.e., create more of them -- which is possible because we can set their
return values). Also, cover the changes made in this commit with tests; and also manually test by
naviagting to /security.txt and observing a redirection to /.well-known/security.txt, which in turn
displays a valid security policy. A status code of 301 was chosen because it is not expected that
the location of the security policy will change... I'm not sure if this is a good idea.

BREAKING CHANGE: Instead of displaying the security policy at /security.txt, redirect (301) to
/.well-known/security.txt and display the policy there

Resolves #57