feat: 301 Redirect from /security.txt to /.well-known/security.txt (#58)
The wording in the specification has changed, so as to encourage redirection to the /.well-known/security.txt location unless there are technical restrictions that prevent that URL from working. As a result, redirect /security.txt to /.well-known/security.txt. Also, change the unit tests a little.

BREAKING CHANGE: Instead of displaying the security policy at /security.txt, 301 redirect to
/.well-known/security.txt and display the policy there

Resolves #57